Creating the next day's forecast

+++ MauiWeather/Models/Daily.cs
@@ -6,10 +6,11 @@ public class Daily
 {
     public string[] Time { get; set; } = [];
     [JsonPropertyName("weather_code")]
-    public int[] WeatherCode { get; set; } = [];
+    public float[] WeatherCode { get; set; } = [];
     [JsonPropertyName("temperature_2m_max")]
     public float[] MaxTemperature { get; set; } = [];
     [JsonPropertyName("temperature_2m_min")]
     public float[] MinTemperature { get; set; } = [];
+
 }
